- Renamed old bls and tests-bls folder for clarity. These will be removed soon. - Updated authors in tests-bls
