+++ old-bls/src/tests/tests_vect_subfunctions.sw
