+++ old-tests-bls/tests/harness.rs
@@ -1,4 +1,4 @@
-mod tests_add_fp;
+// mod tests_add_fp;
 // mod tests_mul_by_x_fp;
 // mod tests_sub_fp;
 // mod tests_l_shift;
@@ -6,4 +6,4 @@ mod tests_add_fp;
 // mod tests_temp_mul;
 // mod tests_vect_subfunctions;
 // mod tests_vect_fp2;
-// mod tests_temp_mul_mont_n; gives error OutOfGas
+mod tests_temp_mul_mont_n;
