+++ test_contract/src/main.sw
@@ -22,7 +22,7 @@ abi EdContract {
     #[storage()]fn scalar_mult(a: Element, x: u32) -> Element;
     #[storage()]fn shift_right_by51(a: U128) -> u64;
 
-    #[storage()]fn inverse(a: Element) -> Element;
+    // #[storage()]fn inverse(a: Element) -> Element;
 
 
 }
@@ -76,7 +76,7 @@ impl EdContract for Contract {
         shift_right_by51(a)
     }
 
-    #[storage()]fn inverse(a: Element) -> Element {
-        inverse(a)
-    }
+    // #[storage()]fn inverse(a: Element) -> Element {
+    //     inverse(a)
+    // }
 }
