+++ tests_bls12_381/Forc.toml
@@ -1,5 +1,5 @@
 [project]
-authors = ["Elena Fuentes Bongenaar"]
+authors = ["Hashcloak"]
 entry = "main.sw"
 license = "Apache-2.0"
 name = "tests_bls12_381"
